requirements.txt
azure-functions==1.13.3
azure-storage-file-share==12.9.0
psycopg==3.1.7
psycopg2-binary==2.9.3
SQLAlchemy==1.4.40
azure-identity==1.11.0              
azure-keyvault-secrets==4.6.0
pytest==7.1.2
pytest-postgresql==4.1.1
alembic==1.7.7
pyjwt==2.6.0
protobuf==3.20.*
pytz==2021.3
marshmallow==3.19.0
apispec==6.3.0
apispec[yaml]